Moved shared labware utility function into protocols/hooks [RCORE-647]
