+++ app/src/organisms/Devices/ProtocolRun/SetupLabware/OffDeckLabwareList.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { SPACING, TYPOGRAPHY } from '@opentrons/components'
 import { StyledText } from '../../../../atoms/text'
 import { LabwareListItem } from './LabwareListItem'
-import type { LabwareSetupItem } from './types'
+import type { LabwareSetupItem } from '../../../../pages/Protocols/hooks'
 
 interface OffDeckLabwareListProps {
   labwareItems: LabwareSetupItem[]
